General Conference Data

A dataset containing all general conference talks back to 1971.

genconf

Format

genconf: A 4-level nested data frame with nestings for conference, session, talk, and paragraph.

genconf A data frame with one row per conference (year + month)

year

Session year

month

Session month

sessions

List dataframe with one row per session.

sessions A data frame one row per session (Saturday AM, PM, etc.)

session_name

individual timepoint

session_id

mean value including imputed values

session_url

Suffix URL path to session (not full url))

talks

List of dataframes, one row per talk in that session

talks A data frame one row per talk

talk_urls

Stub urls for talk.

talk_session_id

Talk index within session

url

Full url path to talk.

title1

Title.

author1

Author Name (typically, might be missing)

author2

Author Role (typically, might be missing)

kicker1

Talk kicker

paragraphs

List of dataframes, one row per talk in that session

paragraphs A data frame one row per paragraph in talk

section_num

If talk has sections, this would be the section number. Newer talks are more likely to have sections.

p_num

Paragraph number

p_id

Paragraph html tag (can be used to generate a url deep link). Might not be in order with p_num due to edge-case talks that use #p1-#p4 for title, author, kicker, etc.

is_header

If a talk contains sections, those sections have headers. Header content will be a few words.

paragraph

Text of talk. <sup></sup> html tags (superscripts/footnotes) have been stripped out.

Source

https://www.churchofjesuschrist.org/study/general-conference