Main function to scrape all conference talk urls For a given year-month conference, return a nested tibble of all sessions with a tibble-column containing the dataframes

Main function to scrape all conference talk urls For a given year-month conference, return a nested tibble of all sessions with a tibble-column containing the dataframes

scrape_conference_urls(year, month)

Arguments

year
year

month
month

Value

tibble

Examples

scrape_conference_urls(2019, 10)

#> # A tibble: 1 × 3
#>    year month sessions        
#>   <dbl> <dbl> <list>          
#> 1  2019    10 <tibble [5 × 4]>

scrape_conference_urls(1971, 4)

#> # A tibble: 1 × 3
#>    year month sessions        
#>   <dbl> <dbl> <list>          
#> 1  1971     4 <tibble [7 × 4]>